refractor: move game class into game_logiv.py
