+++ main.py
@@ -1,0 +1,7 @@
+from game_logic import Game
+
+if __name__ == "__main__":
+    game = Game()
+
+    game.main_loop()
+
